% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acstables.R
\docType{data}
\name{acsagetable}
\alias{acsagetable}
\title{Demographic and sex by age data from ACS tables for counties, states, and 
the U.S. as a whole}
\format{A data frame with 163,600 rows and 8 variables:
\describe{
\item{sex}{Male or female.}
\item{age}{Under 18 years, 18 to 24 years, 25 to 44 years, 45 to 64 years,  
or 65 years and over.}
\item{raceethnicity}{White alone, Hispanic or Latino, black alone, Asian 
alone, or other.}
\item{population}{The number of people in this demographic bin.}
\item{sextotal}{The number of people in this sex and age bin.}
\item{geototal}{The total number of people tabulated in this geographic 
region.}
\item{prob}{\code{population/geototal}, or the proportion of the total that
is in this demographic bin.}
\item{region}{Geographic region, such as \code{US} for the entire United 
States, a two-letter abbrevation for a state (\code{TX} for Texas), or a 
5-digit FIPS code for a county (\code{17031} for Cook County).}
}}
\usage{
acsagetable
}
\description{
Demographic and sex by age data from ACS tables B01001 and B01001B/D/H/I for 
counties, states, and the U.S., fetched and processed by 
\code{process_acs_age} and \code{process_acs_age_all}.
}
\details{
Uses ACS 5-year estimate for 2010-2014; the data from table B01001 
is rebinned and used to find the "other" population
}
\keyword{datasets}

