% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acstables.R
\docType{data}
\name{acsedutable}
\alias{acsedutable}
\title{Demographic and educational attainment data from ACS tables for the states 
and the U.S. as a whole}
\format{A data frame with 2080 rows and 8 variables:
\describe{
\item{sex}{Male or female.}
\item{education}{Less than high school diploma, high school graduate 
(includes equivalency), some college or associate's degree, bachelor's degree 
or higher.}
\item{raceethnicity}{White alone, Hispanic or Latino, black alone, Asian 
alone, or other.}
\item{population}{The number of people in this demographic bin.}
\item{sextotal}{The number of people in this sex and age bin.}
\item{geototal}{The total number of people tabulated in this geographic region.}
\item{prob}{\code{population/geototal}, or the proportion of the total that
is in this demographic bin.}
\item{region}{Geographic region, such as "US" for the entire United States or
a two-letter abbrevation for a state ("TX" for Texas)}
}}
\usage{
acsedutable
}
\description{
Demographic and educational attainment data from ACS tables C15002 and 
C15002B/D/H/I for the states and the U.S., fetched and processed by 
\code{process_acs_education}. The educational attainment tables include only
the population 25 and over.
}
\details{
Uses ACS 1-year estimate for 2014; the data from table C15002 is 
rebinned and used to find the "other" population.
}
\keyword{datasets}

