% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_equivalence.R
\name{baseline_equivalence}
\alias{baseline_equivalence}
\title{Calculate whether matched measures meet WWC standards for baseline equivalence}
\usage{
baseline_equivalence(psm, plot = TRUE)
}
\arguments{
\item{psm}{Output from MatchIt}

\item{plot}{Logical. Whether to include a plot or not.}
}
\value{
A data frame with columns variable, type of variable, the equivalence measure value, whether baseline equivalence was met, and the baseline equivalence measure used.
}
\description{
Calculate whether matched measures meet WWC standards for baseline equivalence
}
\examples{
library(MatchIt); data("lalonde", package = "cobalt")
m.out1 <- matchit(treat ~ age + educ + race +
married + nodegree + re74 + re75,
data = lalonde, method = "nearest")
baseline_equiv <- baseline_equivalence(psm = m.out1)
}
