% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_best_weighting.R
\name{choose_best_weighting}
\alias{choose_best_weighting}
\alias{choose_best_weighting_}
\title{Use bootstrap resampling of survey to find best weighting indicators}
\usage{
choose_best_weighting(mysurvey, n, response, ...)

choose_best_weighting_(mysurvey, n, response_col, dots)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{n}{Number of bootstrap resamplings to generate in order to find the
best weighting indicators}

\item{response}{A column in \code{mysurvey} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{...}{Weighting indicator(s) to be tested for post-stratification.
One or more of \code{sex}, \code{raceethnicity}, \code{age}, and/or 
\code{education} which must be columns in the survey data frame. Both 
\code{age} and \code{education} cannot be used for post-stratification at the
same time because of how the ACS tables are organized.}

\item{response_col}{Response column as string}

\item{dots}{List of weighting indicator(s) as string(s)}
}
\value{
??? WHAT WILL IT RETURN ???
}
\description{
Use bootstrap resampling of survey to find best weighting indicators
}
\details{
\code{choose_best_weighting} is given bare names while 
\code{choose_best_weighting_} is given strings and is therefore suitable for 
programming with. One column of \code{mysurvey} must be \code{geography}, to 
indicate what ACS data to use for post-stratification weighting.
}
\examples{
data(texassurvey)
choose_best_weighting(texassurvey, 10, response, sex, raceethnicity)
data(twostatessurvey)
choose_best_weighting(twostatessurvey, 10, response, sex, education)

}

