% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxs_index.R
\name{coxs_index}
\alias{coxs_index}
\title{Calculate Cox's Index for binary measures}
\usage{
coxs_index(
  prevalance_int = 0.4,
  prevalance_comp = 0.3,
  n_int = 100,
  n_comp = 110
)
}
\arguments{
\item{prevalance_int}{The prevalence of the outcome for the intervention group.}

\item{prevalance_comp}{The prevalence of the outcome for the comparison group.}

\item{n_int}{The number of observations in the intervention group.}

\item{n_comp}{The number of observations in the comparison group.}
}
\value{
The Cox's Index value.
}
\description{
Calculate Cox's Index for binary measures
}
\examples{
i <- coxs_index(prevalance_int = 0.4, prevalance_comp = 0.3, n_int = 100, n_comp = 110)
}
