% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedges_g.R
\name{hedges_g}
\alias{hedges_g}
\title{Calculate Hedges' g for continuous measures}
\usage{
hedges_g(
  int_adj_mean = 1,
  comp_adj_mean = 0.75,
  n_int = 100,
  n_comp = 150,
  int_unadj_sd = 0.1,
  comp_unadj_sd = 0.2
)
}
\arguments{
\item{int_adj_mean}{The intervention group's adjusted mean.}

\item{comp_adj_mean}{The comparison group's adjustment mean.}

\item{n_int}{The number of observations in the intervention group.}

\item{n_comp}{The number of observations in the comparison group.}

\item{int_unadj_sd}{The unadjusted standard deviation of the intervention group.}

\item{comp_unadj_sd}{The unadjusted standard deviation of the comparison group.}
}
\value{
The Hedges' g value.
}
\description{
Calculate Hedges' g for continuous measures
}
\examples{
g <- hedges_g(int_adj_mean = 1.0,
      comp_adj_mean = 0.75,
      n_int = 100,
      n_comp = 150,
      int_unadj_sd = 0.1,
      comp_unadj_sd = 0.2)
}
