% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_of_error.R
\name{margin_of_error}
\alias{margin_of_error}
\alias{margin_of_error_}
\title{Calculate some estimates of the margin of error for a weighted survey 
data frame}
\usage{
margin_of_error(mysurvey, response)

margin_of_error_(mysurvey, response_col)
}
\arguments{
\item{mysurvey}{A survey data frame with post-stratification weights, such 
as those created by \code{weight_wwc}. There must be a column \code{weight} 
in the data frame.}

\item{response}{A column in \code{mysurvey} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{response_col}{Response column as string}
}
\value{
A list with ?????
}
\description{
Calculate some estimates of the margin of error for a weighted survey 
data frame
}
\details{
\code{margin_of_error} is given bare names while 
\code{margin_of_error_} is given strings and is therefore suitable for 
programming with.
}
\examples{
library(dplyr) 
data(texassurvey)
weight_wwc(texassurvey, sex, raceethnicity) \%>\%
    margin_of_error(response)

}

