% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_acs_edu_all.R
\name{process_acs_edu_all}
\alias{process_acs_edu_all}
\title{Fetch and process educational attainment and demographic data from ACS 
tables B15002 and C15002B/D/H/I and return a tidy data frame for a specified
geography with FIPS codes}
\usage{
process_acs_edu_all(geographyfetch, yearspan = 1)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package, such as all counties in the United States}

\item{yearspan}{The span in years of the desired ACS data (should be 1 or
5). Default is 1. Not all county data is available as one-year estimates.}
}
\value{
A data frame with 8 columns that tabulates the sex by age population
for five racial/ethnic groups: black alone, white alone (not Hispanic or 
Latino), Hispanic or Latino, Asian alone, and other. The educational 
attainment binning in this data frame is the same as in ACS tables 
C15002B/D/H/I (the data from table B15002 is rebinned and used to find the 
"other" population).
}
\description{
Fetch and process educational attainment and demographic data from ACS 
tables B15002 and C15002B/D/H/I and return a tidy data frame for a specified
geography with FIPS codes
}
\details{
Uses ACS 1-year estimate for 2014 by default; uses ACS 5-year 
estimate for 2010-2014 with \code{yearspan = 5}. (Not all rural counties
have 1-year estimates.)
}
\examples{
\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
allcounties <- geo.make(state = "*", county = "*")
countiesDF <- process_acs_edu_all(allcounties, yearspan = 5)
}

}

