% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response_age_continuous.R
\name{simulate_response_age_continuous}
\alias{simulate_response_age_continuous}
\title{Simulate the true response to a continuous numerical response question in a 
given geography with specified characteristics for sex, race/ethnicity, 
and age}
\usage{
simulate_response_age_continuous(geographyfetch, lambda_sex,
  lambda_raceethnicity, lambda_age)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}

\item{lambda_sex}{Numeric vector specifying lambda (Poisson distribution) 
for the survey respondents by sex in the order male, then female, for 
example, \code{c(25, 75)}.}

\item{lambda_raceethnicity}{Numeric vector specifying lambda for the survey 
respondents by race/ethnicity in the order white alone, Hispanic or Latino, 
black alone, Asian alone, and other, for example, 
\code{c(90, 10, 50, 50, 50)}.}

\item{lambda_age}{Numeric vector specifying lambda for survey respondents by 
age in the following bins: under 5 years, 5 to 9 years, 10 to 14 years, 15 
to 17 years, 18 and 19 years, 20 to 24 years, 25 to 29 years, 30 to 34 years, 
35 to 44 years, 45 to 54 years, 55 to 64 years, 65 to 74 years, 75 to 84 
years, 85 years and over, for example, 
\code{c(50, 50, 50, 50, 50, 50, 50, 55, 55, 55, 65, 75, 85, 95)}.}
}
\value{
A data frame with 3 columns (\code{value}, \code{answer}, and 
\code{result}) that tabulates the true response to the continuous, numerical 
survey question in the given geography.
}
\description{
Simulate the true response to a continuous numerical response question in a 
given geography with specified characteristics for sex, race/ethnicity, 
and age
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
unitedstates <- geo.make(us = TRUE)
# lambda_sex specifies the response weights of men/women
# in this example, women have a higher mean response than men
lambda_sex <- c(25, 75)
lambda_raceethnicity <- c(90, 10, 50, 50, 50)
lambda_age <- c(50, 50, 50, 50, 50, 50, 50, 55, 55, 60, 65, 70, 75, 80)
opinionDF <- simulate_response_age_continuous(unitedstates, 
                                             lambda_sex, 
                                             lambda_raceethnicity,
                                             lambda_age)
}

}

