% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survey.R
\name{simulate_survey}
\alias{simulate_survey}
\title{Simulate a single survey to a yes/no response question according to inputs}
\usage{
simulate_survey(prop_sex, odds_sex, prop_raceethnicity, odds_raceethnicity,
  prop_age, odds_age, prop_education, odds_education, prop_geography,
  odds_geography, n = 1000)
}
\arguments{
\item{prop_sex}{Numeric vector specifying the gender characteristics of the
survey respondents as proportions in the order male, then female, for
example, \code{c(0.49, 0.51)}. Must sum to 1, i.e., all respondents fall into
one of these gender bins.}

\item{odds_sex}{Numeric vector specifying the opinion odds of the survey
respondents by sex in the order male, then female. For example, \code{c(0.8,
1.25)} means that men are 0.8 times as likely to approve the survey question
and women are 1.25 times as likely to approve the survey question.}

\item{prop_raceethnicity}{Numeric vector specifying the racial/ethnic 
characteristics of the survey respondendents as proportions in the order
white alone, Hispanic or Latino, black alone, Asian alone, and other, 
for example, \code{c(0.6, 0.22, 0.11, 0.05, 0.02)}. Must sum to 1, i.e., all
respondents fall into one of these racial/ethnic bins.}

\item{odds_raceethnicity}{Numeric vector specifying the opinion odds of
the survey respondents by race/ethnicity in the same order as 
\code{prop_raceethnicity}, for example, \code{c(0.2, 2, 2.5, 1, 1)}.}

\item{prop_age}{Numeric vector specifying the age characteristics of the
survey respondents as proportions in the following bins: under 5 years, 5 to 
9 years, 10 to 14 years, 15 to 17 years, 18 and 19 years, 20 to 24 years, 25 
to 29 years, 30 to 34 years, 35 to 44 years, 45 to 54 years, 55 to 64 years, 
65 to 74 years, 75 to 84 years, 85 years and over, for example, \code{c(0, 
0, 0, 0.05, 0.1, 0.1, 0.12, 0.13, 0.12, 0.11, 0.1, 0.1, 0.05, 0.02)}. Must
sum to 1, i.e., all respondents must fall into one of these age bins.}

\item{odds_age}{Numeric vector specifying the opinion odds of the
survey respondents by age in the same order bins as \code{prop_age}, for
example, \code{c(1, 1, 1, 1, 1, 1, 1, 0.5, 2, 2, 2, 2.5, 0.5, 0.2)}.}

\item{prop_education}{Numeric vector specifying the educational attainment
of the survey respondents as proportions in the following bins: less than 
high school diploma, high school graduate (includes equivalency), some 
college or associate's degree, bachelor's degree or higher, for example,
\code{c(0.1, 0.2, 0.4, 0.3)}. Must sum to 1, i.e., all respondents must fall
into one of these educational attainment bins.}

\item{odds_education}{Numeric vector specifying the opinion odds of the survey
respondents by educational attainment in the same order bins as 
\code{prop_education}, for example, \code{c(0.4, 0.5, 2, 2.5)}.}

\item{prop_geography}{Numeric vector specifying the geography distribution
of the survey respondents as proportions in the following bins: Texas, 
California, Utah, for example,
\code{c(0.5, 0, 0.5)}. Must sum to 1, i.e., all respondents must fall
into one of these geography bins.}

\item{odds_geography}{Numeric vector specifying the opinion odds of the 
survey respondents by geography in the same order bins as 
\code{prop_geography}, for example, \code{c(2, 1, 0.5)}.}

\item{n}{Number of respondents in the survey (default is 1000)}
}
\description{
Simulate a single survey to a yes/no response question according to inputs
}
\examples{

# prop_sex specifies how many men/women are in the survey
# in this example, the survey is 48\% men and 52\% women
prop_sex <- c(0.48, 0.52)
# odds_sex specifies the opinions of men/women
# in this example, women are twice as likely to approve and men half as likely
odds_sex <- c(0.5, 2)
prop_raceethnicity <- c(0.55, 0.25, 0.1, 0.05, 0.05)
odds_raceethnicity <- c(0.2, 2, 2.5, 1, 1)
prop_age <- c(0, 0, 0, 0.04, 0.1, 0.1, 0.12, 0.13, 0.12, 0.11, 0.11, 0.09, 0.06, 0.02)
odds_age <- c(1, 1, 1, 1, 1, 1, 1, 0.8, 2, 2, 2.5, 3, 0.5, 0.2)
prop_education <- c(0.1, 0.3, 0.4, 0.2)
odds_education <- c(0.4, 0.5, 2, 2.5)
prop_geography <- c(0.4, 0.3, 0.3)
odds_geography <- c(2, 1, 0.5)
mysurvey <- simulate_survey(prop_sex, odds_sex,
                             prop_raceethnicity, odds_raceethnicity,
                             prop_age, odds_age,
                             prop_education, odds_education,
                             prop_geography, odds_geography,
                             n = 900)

}

