% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_true_response.R
\name{simulate_true_response}
\alias{simulate_true_response}
\title{Simulate the true response to a yes/no response question in a given
geography}
\usage{
simulate_true_response(geographyfetch)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}
}
\value{
A data frame with 3 columns (\code{response}, \code{answer}, and 
\code{result}) that tabulates the true opinion on the yes/no question 
in the given geography.
}
\description{
Simulate the true response to a yes/no response question in a given
geography
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
unitedstates <- geo.make(us = TRUE)
opinionDF <- simulate_true_response(unitedstates)
}

}

