% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_true_response_education.R
\name{simulate_true_response_education}
\alias{simulate_true_response_education}
\title{Simulate the true response to a yes/no response question in a given
population with specified opinions for sex, race/ethnicity, age, and 
geography}
\usage{
simulate_true_response_education(geovector, odds_geography, odds_sex,
  odds_raceethnicity, odds_education)
}
\arguments{
\item{geovector}{A vector of geographies specified as a two letter 
abbreviation or a 5-digit FIPS code. Each geography must be the entire U.S. 
(\code{US}) or a single state (for example, \code{TX} or \code{CA}) or a 
single county (for example, \code{49035}). For example, \code{c("TX", "UT")}.}

\item{odds_geography}{Numeric vector specifying the opinion odds of the 
survey respondents by geography in the same bins as \code{geovector}.}

\item{odds_sex}{Numeric vector specifying the opinion odds of the survey
respondents by sex in the order male, then female. For example, \code{c(0.8,
1.25)} means that men are 0.8 times as likely to approve the survey question
and women are 1.25 times as likely to approve the survey question.}

\item{odds_raceethnicity}{Numeric vector specifying the opinion odds of
the survey respondents by race/ethnicity in the order white alone, Hispanic 
or Latino, black alone, Asian alone, and other, for example, 
\code{c(0.2, 2, 2.5, 1, 1)}.}

\item{odds_education}{Numeric vector specifying the opinion odds of the 
survey respondents by educational attainment in the following bins: less than 
high school diploma, high school graduate (includes equivalency), some 
college or associate's degree, bachelor's degree or higher, for example, 
\code{c(0.4, 0.5, 2, 2.5)}.}
}
\value{
A data frame with 3 columns (\code{value}, \code{answer}, and 
\code{result}) that tabulates the true opinion on the yes/no question 
in the given population.
}
\description{
Simulate the true response to a yes/no response question in a given
population with specified opinions for sex, race/ethnicity, age, and 
geography
}
\examples{
geovector <- c("AL", "CT")
odds_geography <- c(1.5, 0.8)
odds_sex <- c(0.5, 2)
odds_raceethnicity <- c(0.2, 2, 2.5, 1, 1)
odds_education <- c(0.4, 0.5, 2, 2.5)
opinionDF <- simulate_true_response_education(geovector,
                                     odds_geography, 
                                     odds_sex, 
                                     odds_raceethnicity,
                                     odds_education)

}

