% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_survey.R
\name{summarize_survey}
\alias{summarize_survey}
\alias{summarize_survey_}
\title{Summarize a survey data frame}
\usage{
summarize_survey(mysurvey, response)

summarize_survey_(mysurvey, response_col)
}
\arguments{
\item{mysurvey}{A survey data frame with post-stratification weights, such 
as those created by \code{weight_age} or \code{weight_education}. There must
be a column \code{weight} in the data frame.}

\item{response}{A column in \code{mysurvey} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{response_col}{Response column as string}
}
\value{
A data frame with 3 columns (\code{answer}, \code{value}, and 
\code{result}) that tabulates the weighted response on the yes/no question 
in the given geography.
}
\description{
Summarize a survey data frame
}
\details{
\code{summarize_survey} is given bare names while 
\code{summarize_survey_} is given strings and is therefore suitable for 
programming with.
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
texas <- geo.make(state = "TX")
data(texassurvey)
summarize_survey(weight_age(texassurvey, texas, sex, raceethnicity), response)
}

}

