% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesurveys.R
\docType{data}
\name{surveymissing}
\alias{surveymissing}
\title{A simulated survey with missing data}
\format{A data frame with 1000 rows and 5 variables:
\describe{
\item{sex}{Either male or female}
\item{raceethnicity}{White alone (not Hispanic or 
Latino), Hispanic or Latino, black alone, Asian alone, or other}
\item{age}{Under 18 years (no children actually in survey), 18 to 24 years, 
25 to 44 years, 45 to 64 years, or 65 years and over}
\item{education}{Less than high school diploma, high school graduate 
(includes equivalency), some college or associate's degree, bachelor's 
degree or higher}
\item{geography}{\code{TX} only}
\item{response}{Yes or no}
}}
\usage{
surveymissing
}
\description{
A simulated survey created using the \code{simulate_survey} function with 
1000 respondents and about 20% missing data (\code{NA}s).
}
\keyword{datasets}

