% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesurveys.R
\docType{data}
\name{tinysurvey}
\alias{tinysurvey}
\title{A simulated survey with 20 respondents}
\format{A data frame with 20 rows and 5 variables:
\describe{
\item{sex}{Either male or female}
\item{raceethnicity}{White alone (not Hispanic or 
Latino), Hispanic or Latino, black alone, Asian alone, or other}
\item{age}{Under 18 years, 18 to 24 years, 25 to 44 years, 45 to 64 years,  
or 65 years and over}
\item{education}{Less than high school diploma, high school graduate 
(includes equivalency), some college or associate's degree, bachelor's 
degree or higher}
\item{geography}{\code{UT} only}
\item{response}{A numerical, continuous response to a survey question}
}}
\usage{
tinysurvey
}
\description{
A very small simulated survey created using the 
\code{simulate_survey_continuous} function, for use in unit tests, etc.
}
\keyword{datasets}

