% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_age_one.R
\name{weight_age_one}
\alias{weight_age_one}
\title{Weight a survey using post-stratification for a single weighting indicator}
\usage{
weight_age_one(mysurvey, response, indicator, geographyfetch)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{response}{A column in \code{mysurvey} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{indicator}{A weighting indicator to be used for post-stratification.
One of \code{sex}, \code{raceethnicity}, \code{age}}

\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}
}
\value{
A data frame with some columns that have some stuff in them.
}
\description{
Weight a survey using post-stratification for a single weighting indicator
}

