% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin_of_error.R
\name{margin_of_error}
\alias{margin_of_error}
\alias{margin_of_error_}
\title{Calculate some estimates of the margin of error for weighted survey 
results}
\usage{
margin_of_error(inputfile, response)

margin_of_error_(inputfile, response_col)
}
\arguments{
\item{inputfile}{A file containing the weighted survey data (a survey data 
frame with post-stratification weights, such as those created by 
\code{weight_wwc}). There must be a column \code{weight} in the file. Either 
a path to a file, a connection, or literal data (either a single string or a 
raw vector). Files starting with \code{http://}, \code{https://}, 
\code{ftp://}, or \code{ftps://} will be automatically downloaded, and 
zipped files wll be uncompressed.}

\item{response}{A column in \code{inputfile} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{response_col}{Response column as string}
}
\value{
A list with
\itemize{
\item the percent standard error of the raw, unweighted survey,
\item the percent bias of the raw, unweighted survey (compared to the weighted survey),
\item the percent total error of the raw, unweighted survey (the square root of the sum of the squares of the bias and standard error), and
\item the percent standard error of the weighted survey.
}
}
\description{
Calculate some estimates of the margin of error for weighted survey 
results
}
\details{
\code{margin_of_error} is given bare names while 
\code{margin_of_error_} is given strings and is therefore suitable for 
programming with.
}
\examples{
margin_of_error(system.file("extdata/weightedsurvey.csv", package = "WWC"),
    response)

}

