% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_acs_education.R
\name{process_acs_education}
\alias{process_acs_education}
\title{Fetch and process educational attainment and demographic data from ACS 
tables B15002 and C15002B/D/H/I and return a tidy data frame for a specified 
single geography}
\usage{
process_acs_education(geographyfetch, yearspan = 1)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}

\item{yearspan}{The span in years of the desired ACS data (should be 1 or
5). Default is 1. Not all county data is available as one-year estimates.}
}
\value{
A data frame with 7 columns that tabulates the educational attainment
by sex for five racial/ethnic groups: black alone, white alone (not Hispanic 
or Latino), Hispanic or Latino, Asian alone, and other. The educational 
attainment binning in this data frame is the same as in ACS tables 
C15002B/D/H/I (the data from table B15002 is rebinned and used to find the 
"other" population).
}
\description{
Fetch and process educational attainment and demographic data from ACS 
tables B15002 and C15002B/D/H/I and return a tidy data frame for a specified 
single geography
}
\details{
Uses ACS 1-year estimate for 2014 by default; uses ACS 5-year 
estimate for 2010-2014 with \code{yearspan = 5}.
}
\examples{


\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
unitedstates <- geo.make(us = TRUE)
usDF <- process_acs_education(unitedstates)

texas <- geo.make(state = "TX")
txDF <- process_acs_education(texas)

# Cook County (Chicago) is populous enough that it has 1-year estimates
cookcounty <- geo.make(state = "IL", county = 31)
cookDF <- process_acs_education(cookcounty)

# Daggett County in rural Utah has only 5-year estimates
daggettcounty <- geo.make(state = "UT", county = 9)
daggettDF <- process_acs_education(daggettcounty, yearspan = 5)
}

}

