% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ACS_education.R
\name{process_acs_education}
\alias{process_acs_education}
\title{Fetch and process educational attainment and demographic data from ACS 
tables C15002 and C15002B/D/H/I and return a tidy data frame for a specified 
single geography}
\usage{
process_acs_education(geographyfetch)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}
}
\value{
A data frame with 7 columns that tabulates the educational attainment
by sex for five racial/ethnic groups: black alone, white alone (not Hispanic 
or Latino), Hispanic or Latino, Asian alone, and other. The educational 
attainment binning in this data frame is the same as in ACS tables 
C15002B/D/H/I (the data from table C15002 is rebinned and used to find the 
"other" population).
}
\description{
Fetch and process educational attainment and demographic data from ACS 
tables C15002 and C15002B/D/H/I and return a tidy data frame for a specified 
single geography
}
\examples{


\dontrun{
library(acs)
unitedstates <- geo.make(us = TRUE)
usDF <- process_acs_education(unitedstates)

texas <- geo.make(state = "TX")
txDF <- process_acs_education(texas)

cookcounty <- geo.make(county = 17031)
cookDF <- process_acs_education(cookcounty)
}

}

