% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_response_education_continuous.R
\name{simulate_response_education_continuous}
\alias{simulate_response_education_continuous}
\title{Simulate the true response to a continuous numerical response question in a 
given geography with specified characteristics for sex, race/ethnicity, 
and education}
\usage{
simulate_response_education_continuous(geographyfetch, lambda_sex,
  lambda_raceethnicity, lambda_education)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}

\item{lambda_sex}{Numeric vector specifying lambda (Poisson distribution) 
for the survey respondents by sex in the order male, then female, for 
example, \code{c(25, 75)}.}

\item{lambda_raceethnicity}{Numeric vector specifying lambda for the survey 
respondents by race/ethnicity in the order white alone, Hispanic or Latino, 
black alone, Asian alone, and other, for example, 
\code{c(90, 10, 50, 50, 50)}.}

\item{lambda_education}{Numeric vector specifying lambda for survey 
respondents by educational attainment in the following bins: less than 
high school diploma, high school graduate (includes equivalency), some 
college or associate's degree, bachelor's degree or higher, for example, 
\code{c(20, 40, 60, 80)}.}
}
\value{
A data frame with 3 columns (\code{value}, \code{answer}, and 
\code{result}) that tabulates the true response to the continuous, numerical 
survey question in the given geography.
}
\description{
Simulate the true response to a continuous numerical response question in a 
given geography with specified characteristics for sex, race/ethnicity, 
and education
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
unitedstates <- geo.make(us = TRUE)
# lambda_sex specifies the response weights of men/women
# in this example, women have a higher mean response than men
lambda_sex <- c(25, 75)
lambda_raceethnicity <- c(90, 10, 50, 50, 50)
lambda_education <- c(20, 40, 60, 80)
opinionDF <- simulate_response_education_continuous(unitedstates, 
                                                     lambda_sex, 
                                                     lambda_raceethnicity,
                                                     lambda_education)
}

}

