% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survey.R
\name{simulate_survey}
\alias{simulate_survey}
\title{Simulate a single survey to a yes/no response question in a U.S. state}
\usage{
simulate_survey(n = 1000, state, opinionyes)
}
\arguments{
\item{n}{Number of respondents in the survey (default is 1000)}

\item{state}{Two letter string abbreviation for the state to simulate in, for
example, \code{"TX"}, \code{"UT"}, or \code{"WA"}}

\item{opinionyes}{Numeric vector specifying what proportion of each 
racial/ethnic group approves of the survey question, for example, 
\code{c(0.2, 0.7, 0.5, 0.1, 0.9)}}

\item{raceethnicity}{Numeric vector specifying the racial/ethnic 
characteristics of the survey respondendents as proportions in the order
white alone, all Hispanic, black alone, Asian alone, and other, 
for example, \code{c(0.6, 0.22, 0.11, 0.05, 0.02)}. Must sum to 1, i.e., all
respondents fall into one of these racial/ethnic bins.}

\item{gender}{Numeric vector specifying the gender characteristics of the
survey respondents as proportions in the order male, then female, for
example, \code{c(0.49, 0.51)}. Must sum to 1, i.e., all respondents fall into
one of these gender bins.}
}
\description{
Simulate a single survey to a yes/no response question in a U.S. state
}
\examples{

myRaceEthnicity <- c(0.65, 0.25, 0.1, 0.05, 0.05)
myGender <- c(0.48, 0.52)
myOpinions <- c(0.8, 0.2, 0.9, 0.5, 0.5)
mySurvey <- simulate_survey(900, myraceethnicity, myopinions)


}

