% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_survey.R
\name{simulate_survey}
\alias{simulate_survey}
\title{Simulate a single survey to a yes/no response question according to inputs}
\usage{
simulate_survey(n = 1000, prob_sex, weight_sex, prob_raceethnicity,
  weight_raceethnicity, prob_age, weight_age, prob_education, weight_education)
}
\arguments{
\item{n}{Number of respondents in the survey (default is 1000)}

\item{prob_sex}{Numeric vector specifying the gender characteristics of the
survey respondents as proportions in the order male, then female, for
example, \code{c(0.49, 0.51)}. Must sum to 1, i.e., all respondents fall into
one of these gender bins.}

\item{weight_sex}{Numeric vector specifying the opinion weights of the survey
respondents by sex in the order male, then female. For example, \code{c(0.8,
1.25)} means that men are 0.8 times as likely to approve the survey question
and women are 1.25 times as likely to approve the survey question.}

\item{prob_raceethnicity}{Numeric vector specifying the racial/ethnic 
characteristics of the survey respondendents as proportions in the order
white alone, Hispanic or Latino, black alone, Asian alone, and other, 
for example, \code{c(0.6, 0.22, 0.11, 0.05, 0.02)}. Must sum to 1, i.e., all
respondents fall into one of these racial/ethnic bins.}

\item{weight_raceethnicity}{Numeric vector specifying the opinion weights of
the survey respondents by race/ethnicity in the same order as 
\code{prob_raceethnicity}, for example, \code{c(0.2, 2, 2.5, 1, 1)}.}

\item{prob_age}{Numeric vector specifying the age characteristics of the
survey respondents as proportions in the following bins: under 5 years, 5 to 
9 years, 10 to 14 years, 15 to 17 years, 18 and 19 years, 20 to 24 years, 25 
to 29 years, 30 to 34 years, 35 to 44 years, 45 to 54 years, 55 to 64 years, 
65 to 74 years, 75 to 84 years, 85 years and over, for example, \code{c(0, 
0, 0, 0.05, 0.1, 0.1, 0.12, 0.13, 0.12, 0.11, 0.1, 0.1, 0.05, 0.02)}. Must
sum to 1, i.e., all respondents must fall into one of these age bins.}

\item{weight_age}{Numeric vector specifying the opinion weights of the
survey respondents by age in the same order bins as \code{prob_age}, for
example, \code{c(1, 1, 1, 1, 1, 1, 1, 0.5, 2, 2, 2, 2.5, 0.5, 0.2)}.}

\item{prob_education}{Numeric vector specifying the educational attainment
of the survey respondents as proportions in the following bins: less than 
high school diploma, high school graduate (includes equivalency), some 
college or associate's degree, bachelor's degree or higher, for example,
\code{c(0.1, 0.2, 0.4, 0.3)}. Must sum to 1, i.e., all respondents must fall
into one of these educational attainment bins.}

\item{weight_education}{Numeric vector specifying the opinion weights of the survey
respondents by educational attainment in the same order bins as 
\code{prob_education}, for example, \code{c(0.4, 0.5, 2, 2.5)}.}
}
\description{
Simulate a single survey to a yes/no response question according to inputs
}
\examples{

# prob_sex specifies how many men/women are in the survey
# in this example, the survey is 48\% men and 52\% women
prob_sex <- c(0.48, 0.52)
# weight_sex specifies the opinions of men/women
# in this example, women are twice as likely to approve and men half as likely
weight_sex <- c(0.5, 2)
prob_raceethnicity <- c(0.55, 0.25, 0.1, 0.05, 0.05)
weight_raceethnicity <- c(0.2, 2, 2.5, 1, 1)
prob_age <- c(0, 0, 0, 0.04, 0.1, 0.1, 0.12, 0.13, 0.12, 0.11, 0.11, 0.09, 0.06, 0.02)
weight_age <- c(1, 1, 1, 1, 1, 1, 1, 0.8, 2, 2, 2.5, 3, 0.5, 0.2)
prob_education <- c(0.1, 0.3, 0.4, 0.2)
weight_education <- c(0.4, 0.5, 2, 2.5)
mySurvey <- simulate_survey(900, prob_sex, weight_sex,
                             prob_raceethnicity, weight_raceethnicity,
                             prob_age, weight_age,
                             prob_education, weight_education)

}

