% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_true_response_education.R
\name{simulate_true_response_education}
\alias{simulate_true_response_education}
\title{Simulate the true response to a yes/no response question in a given
geography with specified opinions for sex, race/ethnicity, and education}
\usage{
simulate_true_response_education(geographyfetch, odds_sex, odds_raceethnicity,
  odds_education)
}
\arguments{
\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}

\item{odds_sex}{Numeric vector specifying the opinion odds of the survey
respondents by sex in the order male, then female. For example, \code{c(0.8,
1.25)} means that men are 0.8 times as likely to approve the survey question
and women are 1.25 times as likely to approve the survey question.}

\item{odds_raceethnicity}{Numeric vector specifying the opinion odds of
the survey respondents by race/ethnicity in the order white alone, Hispanic 
or Latino, black alone, Asian alone, and other, for example, 
\code{c(0.2, 2, 2.5, 1, 1)}.}

\item{odds_education}{Numeric vector specifying the opinion odds of the 
survey respondents by educational attainment in the following bins: less than 
high school diploma, high school graduate (includes equivalency), some 
college or associate's degree, bachelor's degree or higher, for example, 
\code{c(0.4, 0.5, 2, 2.5)}.}
}
\value{
A data frame with 3 columns (\code{value}, \code{answer}, and 
\code{result}) that tabulates the true opinion on the yes/no question 
in the given geography.
}
\description{
Simulate the true response to a yes/no response question in a given
geography with specified opinions for sex, race/ethnicity, and education
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
unitedstates <- geo.make(us = TRUE)
# odds_sex specifies the opinions of men/women
# in this example, women are twice as likely to approve and men half as likely
odds_sex <- c(0.5, 2)
odds_raceethnicity <- c(0.2, 2, 2.5, 1, 1)
odds_education <- c(0.4, 0.5, 2, 2.5)
opinionDF <- simulate_true_response_education(unitedstates, 
                                     odds_sex, 
                                     odds_raceethnicity,
                                     odds_education)
}

}

