% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_age.R
\name{weight_age}
\alias{weight_age}
\alias{weight_age_}
\title{Weight a survey using post-stratification for weighting indicator(s) from 
ACS tables with age by sex data}
\usage{
weight_age(mysurvey, ...)

weight_age_(mysurvey, dots)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{...}{Weighting indicator(s) to be used for post-stratification.
One or more of \code{sex}, \code{raceethnicity}, and \code{age}, which must 
be columns in the survey data frame.}

\item{dots}{List of weighting indicator(s) as string(s)}
}
\value{
The original survey data frame with 1 column added, \code{weight},
the post-stratification weight for each row in the survey.
}
\description{
Weight a survey using post-stratification for weighting indicator(s) from 
ACS tables with age by sex data
}
\details{
\code{weight_age} is given bare names while \code{weight_age_} is 
given strings and is therefore suitable for programming with. One column of
\code{mysurvey} must be \code{geography}, to indicate what ACS data to use
for post-stratification weighting.
}
\examples{
data(texassurvey)
weight_age(texassurvey, sex, raceethnicity)

}

