% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_age.R
\name{weight_age}
\alias{weight_age}
\alias{weight_age_}
\title{Weight a survey using post-stratification for weighting indicator(s) from 
ACS tables with age by sex data}
\usage{
weight_age_(mysurvey, geographyfetch, response_col, dots, ...)

weight_age(mysurvey, geographyfetch, response, ...)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}

\item{response_col}{A column in \code{mysurvey} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{dots}{List of weighting indicator(s) as string(s)}

\item{...}{Weighting indicator(s) to be used for post-stratification.
One or more of \code{sex}, \code{raceethnicity}, \code{age}}

\item{response}{Response column as string}
}
\value{
A data frame with 3 columns (\code{answer}, \code{value}, and 
\code{result}) that tabulates the weighted response on the yes/no question 
in the given geography.
}
\description{
Weight a survey using post-stratification for weighting indicator(s) from 
ACS tables with age by sex data
}
\details{
\code{weight_age} is given bare names while \code{weight_age_} is 
given strings and is therefore suitable for programming with.
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
texas <- geo.make(state = "TX")
data(texassurvey)
weight_age(texassurvey, texas, response, sex, raceethnicity)
}

}

