% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_age_one.R
\name{weight_age_one}
\alias{weight_age_one}
\title{Weight a survey using post-stratification for a single weighting indicator 
from ACS tables with age by sex data}
\usage{
weight_age_one(mysurvey, response, indicator, geographyfetch)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{response}{A column in \code{mysurvey} that contains the quantity to be
weighted, such as the response to a yes/no question as in 
\code{simulate_survey}}

\item{indicator}{A weighting indicator to be used for post-stratification.
One of \code{sex}, \code{raceethnicity}, \code{age}}

\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}
}
\value{
A data frame with 3 columns (\code{answer}, \code{value}, and 
\code{result}) that tabulates the weighted response on the yes/no question 
in the given geography.
}
\description{
Weight a survey using post-stratification for a single weighting indicator 
from ACS tables with age by sex data
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
unitedstates <- geo.make(us = TRUE)
# prob_sex specifies how many men/women are in the survey
# in this example, the survey is 48\% men and 52\% women
prob_sex <- c(0.48, 0.52)
# weight_sex specifies the opinions of men/women
# in this example, women are twice as likely to approve and men half as likely
weight_sex <- c(0.5, 2)
prob_raceethnicity <- c(0.55, 0.25, 0.1, 0.05, 0.05)
weight_raceethnicity <- c(0.2, 2, 2.5, 1, 1)
prob_age <- c(0, 0, 0, 0.04, 0.1, 0.1, 0.12, 0.13, 0.12, 0.11, 0.11, 0.09, 0.06, 0.02)
weight_age <- c(1, 1, 1, 1, 1, 1, 1, 0.8, 2, 2, 2.5, 3, 0.5, 0.2)
prob_education <- c(0.1, 0.3, 0.4, 0.2)
weight_education <- c(0.4, 0.5, 2, 2.5)
mysurvey <- simulate_survey(prob_sex, weight_sex,
                             prob_raceethnicity, weight_raceethnicity,
                             prob_age, weight_age,
                             prob_education, weight_education,
                             n = 200)
weight_age_one(mysurvey, response, sex, unitedstates)
}

}

