% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_education.R
\name{weight_education}
\alias{weight_education}
\alias{weight_education_}
\title{Weight a survey using post-stratification for weighting indicator(s) from 
ACS tables with educational attainment data}
\usage{
weight_education(mysurvey, ...)

weight_education_(mysurvey, dots)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{...}{Weighting indicator(s) to be used for post-stratification.
One or more of \code{sex}, \code{raceethnicity}, and \code{education}, which 
must be columns in the survey data frame.}

\item{dots}{List of weighting indicator(s) as string(s).}
}
\value{
The original survey data frame with 1 column added, \code{weight},
the post-stratification weight for each row in the survey.
}
\description{
Weight a survey using post-stratification for weighting indicator(s) from 
ACS tables with educational attainment data
}
\details{
\code{weight_education} is given bare names while 
\code{weight_education_} is given strings and is therefore suitable for 
programming with. One column of \code{mysurvey} must be \code{geography}, to 
indicate what ACS data to use for post-stratification weighting.
}
\examples{
data(texassurvey)
weight_education(texassurvey, sex, raceethnicity)

}

