% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_wwc.R
\name{weight_wwc}
\alias{weight_wwc}
\alias{weight_wwc_}
\title{Weight a survey using post-stratification with weighting indicator(s) from 
ACS tables}
\usage{
weight_wwc(mysurvey, ..., force_edu = FALSE)

weight_wwc_(mysurvey, dots, force_edu = FALSE)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{...}{Weighting indicator(s) to be used for post-stratification.
One or more of \code{sex}, \code{raceethnicity}, \code{age}, and/or 
\code{education} which must be columns in the survey data frame. Both 
\code{age} and \code{education} cannot be used for post-stratification at the
same time because of how the ACS tables are organized.}

\item{force_edu}{Should the weighting be done with the ACS education tables? 
Defaults to \code{FALSE}; cannot be \code{TRUE} if \code{age} is one of the
weighting indicators}

\item{dots}{List of weighting indicator(s) as string(s)}
}
\value{
The original survey data frame with 1 column added, \code{weight},
the post-stratification weight for each row in the survey.
}
\description{
Weight a survey using post-stratification with weighting indicator(s) from 
ACS tables
}
\details{
\code{weight_wwc} is given bare names while \code{weight_wwc_} is 
given strings and is therefore suitable for programming with. One column of 
\code{mysurvey} must be \code{geography}, to indicate what ACS data to use 
for post-stratification weighting.
}
\examples{
data(texassurvey)
weight_wwc(texassurvey, sex, raceethnicity)
data(twostatessurvey)
weight_wwc(twostatessurvey, sex, education)

}

