% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_wwc.R
\name{weight_wwc}
\alias{weight_wwc}
\alias{weight_wwc_}
\title{Weight a survey using post-stratification with weighting indicator(s) from 
ACS tables}
\usage{
weight_wwc(mysurvey, geographyfetch, ...)

weight_wwc_(mysurvey, geographyfetch, dots)
}
\arguments{
\item{mysurvey}{A survey data frame, such as that created by 
\code{simulate_survey}}

\item{geographyfetch}{A geography created with the \code{geo.make()} function
of the acs package. The geography must be the entire U.S. or a single state
or a single county.}

\item{...}{Weighting indicator(s) to be used for post-stratification.
One or more of \code{sex}, \code{raceethnicity}, \code{age}, and/or 
\code{education} which must be columns in the survey data frame. Both 
\code{age} and \code{education} cannot be used for post-stratification at the
same time because of how the ACS tables are organized.}

\item{dots}{List of weighting indicator(s) as string(s)}
}
\value{
The original survey data frame with 1 column added, \code{weight},
the post-stratification weight for each row in the survey.
}
\description{
Weight a survey using post-stratification with weighting indicator(s) from 
ACS tables
}
\details{
\code{weight_wwc} is given bare names while \code{weight_wwc_} is 
given strings and is therefore suitable for programming with.
}
\examples{

\dontrun{
library(acs)
# if you are new to using the acs package, you will need to get an API key
# and run api.key.install() one time to install your key on your system
texas <- geo.make(state = "TX")
data(texassurvey)
weight_wwc(texassurvey, texas, sex, raceethnicity)
}

}

